refactor: reduce code issues
